Created Mars Icon Method
Mars is always sunny, apparently. I couldn't find a single instance of
any other sky condition in the mars weather archive on
marsweather.ingenology.com . Thus, the icon for mars is always sunny,
unless something goes awry in which case it is a thunderstorm icon to
say something is wrong.