Fixed a Bug
For input of city as something other than a city, such as "canada" or
"texas", weather data is retrieved but the city name returned in the
json is "". Fixed so that in this case the weatherdata class will use
the original input to make the other requests, rather than the null
string.