Improvements
Lots of little fixes, many are case specific. For example, if a city is
added to the my locations list, it becomes selected on that list, etc.
:coffee:
Also fixed Time class so that it will be correct with regard to daylight
savings, and updates to the current time upon refreshing the weather
data.